refactor: move const to config
