+++ apps/github/src/config.ts
@@ -2,6 +2,7 @@ import { EaveConfig } from "@eave-fyi/eave-stdlib-ts/src/config.js";
 import { EaveApp } from "@eave-fyi/eave-stdlib-ts/src/eave-origins.js";
 
 export const GITHUB_EVENT_QUEUE_NAME = "github-events-processor";
+export const API_BRANCH_NAME = "refs/heads/eave/auto-docs/api";
 
 class AppConfig extends EaveConfig {
   eaveOrigin = EaveApp.eave_github_app;
