--- apps/github/src/constants.ts
@@ -1,1 +1,0 @@
-export const API_BRANCH_NAME = "refs/heads/eave/auto-docs/api";
